kernelreport.py: fix the testcases number incorrect problem
as the vts-job is always set with the last vts job,
either the v7a vts job or the v8a vts, it makes the tallyNumbers
function only plus the number of one vts job.
The correct one should have the number in both vts jobs.
This change makes the vts-job only be set when there is no vts v7a job
and no vts v8a job, which would help for the check make
calculate correctly

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>