kernelreport.py: fix the problem to get results for x15 builds
changed to use the number of finished successfully jobs for check.

when the finished successfully jobs number is the same as the number of all jobs
it means all the jobs are finished successfully, the build is OK to be used for comparison.
It should has results for all the cts vts jobs

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>